refactor(manifeste): le presse-papiers et les gestes de composition sortent d'app.js
Deux modules de plus, même méthode que les deux précédents.

`presse-papiers.js` — six gestes qui font tous la même chose : mettre un texte
dans le presse-papiers et le dire par le bouton qui l'a demandé. Ils n'ont
jamais eu de rapport avec une vue.

`manifeste-gestes.js` — ajouter une suggestion, ajouter librement une
commodité, retirer une ligne, ajuster les SCU, revenir à l'optimal, et
abandonner la composition quand on change de route. Chacun RECALCULE le
chargement au lieu de le relire d'une globale.

DEUX ERREURS ATTRAPÉES PAR LA SUITE, toutes deux du même genre — un
identifiant oublié dans un déplacement, invisible à `tsc` :

1. `saveState` non importé dans `presse-papiers.js`. « ⟐ Partager » ne copiait
   plus rien et ne le disait pas : le `try/catch` du presse-papiers avalait
   l'exception exactement comme celui de `withMarket` l'avait fait plus tôt
   dans la soirée. Deux tests l'ont vu — celui du lien périmé (#54) et celui du
   nom accessible (#59).
2. `resetManifeste` gardait un `refresh()` : c'est le seul des cinquante gestes
   déplacés dont le corps tient sur UNE ligne, et le remplacement en masse ne
   visait que les lignes indentées. « ↺ optimal » ne rendait plus la main au
   calcul (#19).

Le contrôle systématique des identifiants non résolus, ajouté après la
première, aurait attrapé les deux.

Quatorze imports morts de plus quittent `app.js`.

app.js : 1 670 → 1 550 lignes.
Vérification : 244 e2e — 242 passés, 2 ignorés, 0 échec — · 490/490 unitaires ·
tsc silencieux · build OK.
